refactor(dialog): refactor dialog reference code
We separate the file to give the DialogRef its own file. We also remove some if-else statements. The
reference to the dialog returned by the .open() method is one object that developers can use: it's
also a IDialog that exposes a close() method. No change in the API, it's just easier to read and
maintain now.